requirements.txt
anytree==2.0.0
apache-libcloud==1.1.0
Babel==2.3.4
backports-abc==0.4
certifi==2017.7.27.1
cffi==1.6.0
click==6.6
colorclass==2.2.0
colorlog==2.6.0
coverage==4.3.4
cryptography==1.4
debtcollector==1.12.0
Django==1.9.7
django-configuration==0.0.3
django-polymorphic==0.9.2
ecdsa==0.13
enum34==1.1.6
Flask==0.11.1
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.16.0
futures==3.0.5
hypchat==0.21
idna==2.1
ipaddress==1.0.16
iso8601==0.1.11
itsdangerous==0.24
Jinja2==2.8
jsonpatch==1.15
jsonpath==0.54
jsonpickle==0.9.3
jsonpointer==1.10
jsonschema==2.6.0
keystoneauth1==2.18.0
lxml==3.6.0
MarkupSafe==0.23
monotonic==1.2
msgpack-python==0.4.8
mysql-connector-python==2.1.3
MySQL-python==1.2.5
ndg-httpsclient==0.4.1
netaddr==0.7.19
netifaces==0.10.5
olefile==0.44
oslo.i18n==3.13.0
oslo.serialization==2.17.0
oslo.utils==3.23.0
paramiko==1.17.2
pbr==2.0.0
pex==1.2.11
Pillow==4.2.1
pluggy==0.4.0
positional==1.1.1
prettytable==0.7.2
py==1.4.32
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
pyOpenSSL==16.0.0
pyparsing==2.2.0
pytest==3.0.6
pytest-cov==2.4.0
python-dateutil==2.6.0
python-glanceclient==2.6.0
python-novaclient==7.1.0
python-telegram-bot==8.0
pytz==2016.10
requests==2.13.0
ruamel.ordereddict==0.4.9
ruamel.yaml==0.14.11
semver==2.4.0
simplejson==3.10.0
singledispatch==3.4.0.3
six==1.10.0
stevedore==1.21.0
suds==0.4
telepot==12.3
terminaltables==2.1.0
tornado==4.4
tox==2.7.0
typing==3.5.3.0
urllib3==1.22
virtualenv==15.1.0
warlock==1.2.0
Werkzeug==0.11.10
wrapt==1.10.8
xmltodict==0.10.2
